refactor parser tree
